Allow for setting a tabindex on the date picker
Note that we can't use the `tabindex` attribute since this would actually set the tab index on the component tab when binding to it in the usage. Instead we'll bind in `inputTabIndex` and set this to the tab index of the input component itself, which is what we want it on.