Provide a component method for focusing the element
There are some cases where the consumer needs to programatically focus a particular date input. In order to support this, we need to get a reference to the input element, then use the angular renderer to invoke focus on it. (The renderer allows us to do this platform agnostically, such that this wouldn't break outside of the browser) 

Note that both the mobile and desktop inputs are tagged with the same template ref, so no matter which case is true, the component will focus the input.