Mark years, months, and days as disabled if they fall outside of min-max
This is the most complex behavior for min and max dates. 

For a min date of 6/1/2018, we want all years prior to 2018 to show as disabled. If an earlier year is selected, show all months and days as disabled. If a later year is selected, all months and days should show as enabled. If 2018 is selected, all months prior to 6/June should be disabled, and so on for days.

For a max date of 6/1/2018, all of the same logic is applied in reverse.